test: lock preload daemon channel parity
